Added Cuda Extension, added test for unalinged store
Cuda works with raw pointers. To handle those, a wrapper struct was introduced, which either allocates memory on the device and puts it into a std::shared_ptr (thus, the memory is freed when the register type goes out of scope) or stores the raw pointer (to avoid unnecessary copy operations). 
However, this has to be changed before merging, since on the one hand, std::shared_ptr is not available per default for cuda and on the other hand malloc / free has to be substituted with their cuda-counterparts.