Changed sse3 flag to ssse3 flag
SSSE3 is a superset of SSE3. It seems to be the case, that intel only reports on the availability of SSSE3. Since the generator does not provide any fancy logical operators for the provided lscpu flags, we stick to SSSE3 and provide a specialization comment, indicating what it actually mean.