First try to overcome the overload ambiguity
If an overload for a free function exists, it has to be checked, whether the parameters of the two functions are not equal (otherwise, the overload must be disabled). 
To achieve that, I added a parameter_tuple to every implementation struct which is used for SFNIAE comparison at the free-function level.
However, currently mask_population_count and conflict_free are generating errors.